refactor(RacingGame): Game->RacingGame 클래스명 변경
변경사항
- Game class -> RacingGame class 이름 수정
- allRoundStart 메서드 -> start 메서드 이름 수정
