+++ src/main/java/racingcar/racing/model/RacingGameFactory.java
@@ -3,7 +3,7 @@ package racingcar.racing.model;
 import java.util.List;
 
 public interface RacingGameFactory {
-    Game createGame(int allRound, List<Car> cars);
+    RacingGame createGame(int allRound, List<Car> cars);
 
     List<Car> createCars(List<String> carNames);
 }
