+++ src/test/java/racingcar/racing/controller/InputValidatorTest.java
@@ -4,6 +4,7 @@ import java.util.Arrays;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
+import racingcar.racing.utils.InputValidator;
 
 class InputValidatorTest {
